requirements.txt
docopt==0.6.2
future==0.15.2
python-telegram-bot==3.3
